More class renaming and imports clean up.
