+++ deltakit-explorer/tests/conftest.py
@@ -7,8 +7,12 @@ import numpy as np
 import pytest
 
 from deltakit_explorer._utils._utils import DELTAKIT_SERVER_URL_ENV
-from deltakit_explorer.analysis._lambda import LambdaData
-from deltakit_explorer.analysis._leppr import LogicalErrorProbabilityPerRoundResults
+from deltakit_explorer.analysis import (
+    LambdaData,
+)
+from deltakit_explorer.analysis import (
+    LogicalErrorProbabilityPerRoundData as LEPPRData,
+)
 
 
 def pytest_sessionstart(session):  # noqa: ARG001
@@ -31,8 +35,8 @@ def lambda_results() -> LambdaData:
 
 
 @pytest.fixture
-def leppr_results() -> LogicalErrorProbabilityPerRoundResults:
-    return LogicalErrorProbabilityPerRoundResults(
+def leppr_results() -> LEPPRData:
+    return LEPPRData(
         leppr=0.001,
         leppr_stddev=0.0001,
         spam_error=0.01,
